+++ src/server/physics/event/ig_ENTITY_Combatant.ts
@@ -1,4 +1,5 @@
 import { prestart } from '../../../loading-stages'
+import { assert } from '../../../misc/assert'
 import { PhysicsServer } from '../physics-server'
 import { setNextSetBy, unsetNextSetBy } from './vars'
 
@@ -6,6 +7,10 @@ declare global {
     namespace ig.ENTITY {
         interface Combatant {
             lastDamagedNetid?: string
+            lastDamagedNetidPlayer?: string
+
+            getLastDamagingEntity(this: this): ig.Entity | undefined
+            getLastDamagingPlayer(this: this): dummy.DummyPlayer | undefined
         }
     }
 }
@@ -14,11 +19,25 @@ prestart(() => {
 
     ig.ENTITY.Combatant.inject({
         onDamage(damagingEntity, attackInfo, animPart) {
-            this.lastDamagedNetid = damagingEntity.getCombatantRoot()?.netid
+            const root = damagingEntity.getCombatantRoot()
+            if (root) {
+                this.lastDamagedNetid = root.netid
+                if (root instanceof dummy.DummyPlayer) this.lastDamagedNetidPlayer = root.netid
+            }
             return this.parent(damagingEntity, attackInfo, animPart)
         },
+        getLastDamagingEntity() {
+            if (this.lastDamagedNetid) return ig.game.entitiesByNetid[this.lastDamagedNetid]
+        },
+        getLastDamagingPlayer() {
+            if (this.lastDamagedNetidPlayer) {
+                const player = ig.game.entitiesByNetid[this.lastDamagedNetidPlayer]
+                assert(player instanceof dummy.DummyPlayer)
+                return player
+            }
+        },
         selfDestruct(resolveDefeat) {
-            const entity = this.lastDamagedNetid && ig.game.entitiesByNetid[this.lastDamagedNetid]
+            const entity = this.getLastDamagingEntity()
             if (!(multi.server instanceof PhysicsServer) || !entity) return this.parent(resolveDefeat)
 
             setNextSetBy(entity)
@@ -26,7 +45,7 @@ prestart(() => {
             unsetNextSetBy()
         },
         update() {
-            const entity = this.lastDamagedNetid && ig.game.entitiesByNetid[this.lastDamagedNetid]
+            const entity = this.getLastDamagingEntity()
             if (!(multi.server instanceof PhysicsServer) || !entity) return this.parent()
 
             setNextSetBy(entity)