+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Fix some options weirdness by linking base instance sc.OptionModel to server and map sc.OptionModel
 - Fix some music weirdness by linking base instance ig.Music to server and map ig.Music
+- Update base instance music volume after stopping server
 
 ## [0.5.3] 2025-10-29
 ## [0.5.2] 2025-10-29