+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Fix some options weirdness by linking base instance sc.OptionModel to server and map sc.OptionModel
 - Fix some music weirdness by linking base instance ig.Music to server and map ig.Music
 - Update base instance music volume after stopping server
+- Fix crash when enemy drops item and player quickly leaves map
+- Direct enemy drops to the player that finished it off
+- Fix npc runners spawn rate multiplying with the number of clients on a map
 
 ## [0.5.3] 2025-10-29
 ## [0.5.2] 2025-10-29